\name{RcppGuile-package}
\alias{RcppGuile-package}
\alias{RcppGuile}
\docType{package}
\title{\packageTitle{RcppGuile}}
\description{\packageDescription{RcppGuile}}
\section{Package Content}{\packageIndices{RcppGuile}}
\author{\packageAuthor{RcppGuile}}
\section{Maintainer}{\packageMaintainer{RcppGuile}}
\keyword{package}
